+++ sns_service/src/main/kotlin/joryu/sns_service/comment/dto/request/CommentUpdateRequest.kt
@@ -1,4 +1,4 @@
-package joryu.sns_service.comment.dto.req
+package joryu.sns_service.comment.dto.request
 
 data class CommentUpdateRequest(
     val content: String
