+++ sns_service/src/main/kotlin/joryu/sns_service/post/dto/request/PostUpdateRequest.kt
@@ -1,4 +1,4 @@
-package joryu.sns_service.post.dto.req
+package joryu.sns_service.post.dto.request
 
 data class PostUpdateRequest(
     val content: String
