+++ sns_service/src/main/kotlin/joryu/sns_service/post/dto/response/PostLikeCountResponse.kt
@@ -1,4 +1,4 @@
-package joryu.sns_service.post.dto.resp
+package joryu.sns_service.post.dto.response
 
 data class PostLikeCountResponse(
     val likeCount: Long,
